+++ src/Objects/Foods/Cupcake.cpp
@@ -13,7 +13,7 @@ Cupcake::Cupcake(Team* team, Shop* shop) : Object("cupcake", team, shop, ObjType
 
 Cupcake::~Cupcake() { }
 
-void Cupcake::on_buy(int index) {
+void Cupcake::on_buy(size_t index) {
     spdlog::debug("{} bought !", name);
     std::vector<Pet*>& team_pets = get_team_pets();
